refactored, split logic into differnet files
