+++ src/core.rs
@@ -1,19 +1,16 @@
 use crate::breakpoint::*;
 use crate::functions::*;
 use crate::process::*;
-use capstone::prelude::*;
 use nix::sys::ptrace;
 use nix::sys::ptrace::getregs;
 use nix::sys::ptrace::setregs;
 use nix::sys::wait::{waitpid, WaitStatus};
-use nix::unistd::Pid;
-use std::io;
 use std::path::Path;
 use std::process::Command;
 
-use crate::debugger::command::CommandHandler;
+use crate::command::CommandHandler;
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum DebuggerState {
     Interactive,
     AwaitingTrap,
@@ -39,20 +36,19 @@ impl Debugger {
         }
     }
 
-    fn read_process_memory(&self, pid: Pid, addr: usize, buf: &mut [u8]) -> bool {
-        let local = iovec {
-            iov_base: buf.as_mut_ptr() as *mut _,
-            iov_len: buf.len(),
-        };
-
-        let remote = iovec {
-            iov_base: addr as *mut _,
-            iov_len: buf.len(),
-        };
-
-        let result = unsafe { process_vm_readv(pid.as_raw() as pid_t, &local, 1, &remote, 1, 0) };
-
-        result == buf.len() as isize
+    pub fn run(&mut self) {
+        loop {
+            let state = self.state.clone();
+            match state {
+                DebuggerState::AwaitingTrap => self.resume_and_wait(),
+                DebuggerState::Interactive => {
+                    let mut handler = CommandHandler { debugger: self };
+                    let input = handler.get_command();
+                    handler.handle_command(input.as_str());
+                }
+            }
+            println!("state: {:?}", self.state);
+        }
     }
 
     fn handle_sigtrap(&mut self) {
@@ -64,22 +60,8 @@ impl Debugger {
             self.breakpoint
                 .remove_breakpoint(cur_addr, self.process.pid);
             regs.rip -= 1;
-            setregs(self.process.pid, regs);
+            let _ = setregs(self.process.pid, regs);
             ptrace::step(self.process.pid, None).expect("Single-step after breakpoint failed");
-        }
-        self.state = DebuggerState::Interactive;
-    }
-
-    pub fn run(&mut self) {
-        loop {
-            let mut handler = CommandHandler { debugger: self };
-            match (self.state) {
-                DebuggerState::AwaitingTrap => self.resume_and_wait(),
-                DebuggerState::Interactive => {
-                    let input = handler.get_command();
-                    handler.handle_command(input.as_str());
-                }
-            }
         }
     }
 
@@ -99,6 +81,7 @@ impl Debugger {
                 if signal == nix::sys::signal::Signal::SIGTRAP {
                     self.handle_sigtrap();
                 }
+                self.state = DebuggerState::Interactive;
             }
             Ok(WaitStatus::Signaled(_, signal, _)) => {
                 println!("Process terminated by signal: {:?}", signal);
@@ -118,17 +101,15 @@ impl Debugger {
 }
 
 fn get_pid_from_input(input: String) -> i32 {
-    let mut pid: i32 = 0;
     if Path::new(&format!("/proc/{}", input)).is_dir() {
         println!("{} is a pid", input);
-        pid = input.parse().expect("Failed to parse PID");
+        input.parse().expect("Failed to parse PID")
     } else if Path::new(&input).is_file() {
         println!("{} is a file", input);
         println!("Executing {}", input);
         let child = Command::new(input).spawn().unwrap();
-        pid = child.id() as i32;
+        child.id() as i32
     } else {
         panic!("provided pid|path not valid");
     }
-    return pid;
 }
